#116 add some functions as the library.
